admin_role

+++ routes/web.php
@@ -20,9 +20,9 @@ Route::get('/', function () {
 });
 
 Route::get('/tahfidz', [TahfidzController::class, 'index'])->middleware('auth');
-Route::get('/tahfidz/create', [TahfidzController::class, 'create']);
-Route::post('/tahfidz', [TahfidzController::class, 'store']);
-Route::delete('/tahfidz/{id}', [TahfidzController::class, 'delete']);
-Route::get('/tahfidz/{id}/edit', [TahfidzController::class, 'edit']);
-Route::patch('/tahfidz/{id}', [TahfidzController::class, 'update']);
+Route::get('/tahfidz/create', [TahfidzController::class, 'create'])->middleware('is_admin');
+Route::post('/tahfidz', [TahfidzController::class, 'store'])->middleware('is_admin');
+Route::delete('/tahfidz/{id}', [TahfidzController::class, 'delete'])->middleware('is_admin');
+Route::get('/tahfidz/{id}/edit', [TahfidzController::class, 'edit'])->middleware('is_admin');
+Route::patch('/tahfidz/{id}', [TahfidzController::class, 'update'])->middleware('is_admin');
